pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Self-improving CLI coding agent with LangGraph server"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "langchain-anthropic>1.1.0",
    "langgraph>=1.0.8",
    "langgraph-sdk>=0.1.0",
    "langgraph-cli[inmem]>=0.4.12",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
codie = "agent.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
